+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.18.0] - 2021-12-12
+### Added
+- Previous week reading for electricity week chart.
+
+### Changed
+- Chart legend labels.
+
 ## [1.17.0] - 2021-12-11
 ### Added
 - Previous week reading for gas week chart.